+++ src/main/java/ru/caloricity/carbohydratesresearch/CarbohydratesResearchUpdateDto.java
@@ -3,9 +3,9 @@ package ru.caloricity.carbohydratesresearch;
 import jakarta.validation.constraints.NotNull;
 
 public record CarbohydratesResearchUpdateDto(
-        @NotNull Float byuksaParallelFirst,
-        @NotNull Float byuksaParallelSecond,
-        @NotNull Float byuksaAfterDryingParallelFirst,
-        @NotNull Float byuksaAfterDryingParallelSecond
+        @NotNull Double byuksaParallelFirst,
+        @NotNull Double byuksaParallelSecond,
+        @NotNull Double byuksaAfterDryingParallelFirst,
+        @NotNull Double byuksaAfterDryingParallelSecond
 ) {
 }
