+++ src/main/java/ru/caloricity/proteinsresearch/ProteinsResearchDto.java
@@ -3,9 +3,9 @@ package ru.caloricity.proteinsresearch;
 import java.util.UUID;
 
 public record ProteinsResearchDto(UUID id,
-                                  Float titrantVolumeParallelFirst,
-                                  Float titrantVolumeParallelSecond,
-                                  Float controlVolume,
-                                  Float coefficient
+                                  Double titrantVolumeParallelFirst,
+                                  Double titrantVolumeParallelSecond,
+                                  Double controlVolume,
+                                  Double coefficient
 ) {
 }
